fix: Width-coerce the len() builtin on Rust and Kotlin
- Infer len(bytes|str) as UntypedInt so comparisons adopt the width
- Cast usize Rust .len() to the expected int width at emit time
- Kotlin .size gains .toUInt(); bless 2 goldens, rest unchanged

+++ tests/forge/expected/algorithm_bytes_equal.rs
@@ -16,7 +16,7 @@ pub fn bytes_equal(a: &[u8], b: &[u8]) -> bool {
         return false;
     }
     let mut i: u32 = 0;
-    while i < (a).len() {
+    while i < (a).len() as u32 {
         if a[(i) as usize] != b[(i) as usize] {
             return false;
         }
